+++ src/modules/vehicle/models/vehicle-model.ts
