+++ src/shared/connection/typeorm/index.ts
@@ -1,6 +1,6 @@
 import { DataSource } from 'typeorm';
-import { User } from '../../../modules/users/models/User';
-import { Vehicle } from '../../../modules/vehicle/models/Vehicle';
+import { User } from '../../../modules/users/models/user-model';
+import { Vehicle } from '../../../modules/vehicle/models/vehicle-model';
 
 import { createUsers1656982804350 } from '../migrations/1656982804350-create-users';
 import { createVehicle1656988043492 } from '../migrations/1656988043492-create-vehicle';
