--- src/shared/container/index.ts
@@ -1,10 +1,0 @@
-import { container } from 'tsyringe';
-import { UsersRepository } from '../../modules/users/repositories/implementations/UsersRepository';
-import { IUsersRepository } from '../../modules/users/repositories/IUsersRepository';
-
-import '../providers';
-
-container.registerSingleton<IUsersRepository>(
-  'UsersRepository',
-  UsersRepository,
-);
