--- src/shared/providers/index.ts
@@ -1,8 +1,0 @@
-import { container } from 'tsyringe';
-import { IBcryptHashProvider } from './bcrypt/IBcryptHashProvider';
-import { BcryptHashProvider } from './bcrypt/implementations/BcryptHashProvider';
-
-container.registerSingleton<IBcryptHashProvider>(
-  'BcryptHashProvider',
-  BcryptHashProvider,
-);
